Files moved to directories. Small changes
