+++ LemonUI.SHV.Example/src/dllmain.cpp
@@ -1,4 +1,4 @@
-#include "pch.h"
+#include "pch.hpp"
 #include "Example.hpp"
 
 #include <Helpers.hpp>
